pyproject.toml
[project]
name = "bangkok-taxi-ops"
version = "0.1.0"
description = "End-to-End MLOps Pipeline for Bangkok Taxi Revenue Optimization"
authors = [
    {name = "Papangkorn Wongsuwan",email = "101549429+skywalker-89@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "xgboost (>=3.1.2,<4.0.0)",
    "mlflow (==2.12.1)",
    "prefect (>=3.6.4,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "h3 (>=4.3.1,<5.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

